PR-310 Заменить IntegerPercent на double

+++ app/src/main/java/com/dmi/perfectreader/menu/MenuPresenter.java
@@ -23,6 +23,6 @@ public class MenuPresenter extends BasePresenter {
     }
 
     public void requestCurrentPercent(int maxPosition) {
-        view.setPosition(bookPresenter.currentPercent() * maxPosition);
+        view.setPosition((int) (bookPresenter.currentPercent() * maxPosition));
     }
 }
